requirements.txt
# Core dependencies
streamlit>=1.33.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-qdrant>=0.1.0
langchain-cohere>=0.1.0
langgraph>=0.1.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0

# Optional for enhanced functionality
cohere>=4.0.0
tavily-python>=0.3.0